Move the visual editor into the tools/editor crate
The slint-editor binary becomes its own unpublished crate instead of
an example of slint-lsp: it no longer compiles the language server and
formatter, and the sparklers auto-updater (a git dependency, which
cargo publish rejects) moves with it, keeping slint-lsp publishable.

The packaging files and every workflow, script, doc, and config
reference follow the binary: artifact paths lose the examples/
segment, version lookups key on the slint-editor package, and the
macOS packaging script picks up the example sources from
internal/editor-preview, where the UI now lives.

The editor's builds intentionally drop the preview-remote and
preview-external features: the editor window has no remote-preview UI
and ignored those messages already, so this only stops linking mdns-sd
and the WebSocket stack into it.
