+++ .github/workflows/visual_editor_macos_dmg.yaml
@@ -105,9 +105,9 @@ jobs:
 
       - name: Compute DMG name
         run: |
-          version="$(cargo metadata --format-version 1 --no-deps | jq -r 'first(.packages[] | select(.name == "slint-lsp") | .version)')"
+          version="$(cargo metadata --format-version 1 --no-deps | jq -r 'first(.packages[] | select(.name == "slint-editor") | .version)')"
           if [ -z "$version" ] || [ "$version" = "null" ]; then
-              echo "Could not determine slint-lsp version from Cargo metadata" >&2
+              echo "Could not determine slint-editor version from Cargo metadata" >&2
               exit 1
           fi
           echo "SLINT_DMG_NAME=SlintVisualEditor-${version}-macos-arm64" >> "$GITHUB_ENV"
@@ -118,7 +118,7 @@ jobs:
         with:
           dmg_name: Slint Visual Editor
           src_dir: target/macos-visual-editor-dmg/dmg-stage/Slint Visual Editor.app
-          background: tools/lsp/packaging/macos/dmg-background.svg
+          background: tools/editor/packaging/macos/dmg-background.svg
           window_size: 660x420
           icon_size: "112"
           icon_position: 184 278
