+++ .gitignore
@@ -18,10 +18,10 @@ Pipfile.lock
 /dist
 
 # Generated by XcodeGen for the macOS visual editor package.
-/tools/lsp/Info.plist
-/tools/lsp/dist
-/tools/lsp/Slint Visual Editor.xcodeproj
-/tools/lsp/packaging/macos/dmg-background.png
+/tools/editor/Info.plist
+/tools/editor/dist
+/tools/editor/Slint Visual Editor.xcodeproj
+/tools/editor/packaging/macos/dmg-background.png
 
 # Ignore all package-lock.json files
 **/package-lock.json
@@ -61,6 +61,6 @@ MODULE.bazel.lock
 Info.plist
 
 # Generated by scripts/build_visual_editor_flatpak.bash
-tools/lsp/cargo-sources.json
-tools/lsp/skia-sources.json
+tools/editor/cargo-sources.json
+tools/editor/skia-sources.json
 .flatpak-builder/
