+++ REUSE.toml
@@ -206,13 +206,13 @@ path = [
   "tools/viewer/android/fastlane/metadata/android/**.png",
 
   # Visual editor-related files
-  "tools/lsp/packaging/linux/**",
-  "tools/lsp/packaging/macos/AppIcon.icon/Assets/icon_backdrop.svg",
-  "tools/lsp/packaging/macos/AppIcon.icon/icon.json",
-  "tools/lsp/packaging/macos/dmg-background.svg",
-  "tools/lsp/packaging/tile-backdrop-wide.svg",
-  "tools/lsp/packaging/tile-backdrop.svg",
-  "tools/lsp/packaging/windows-appx/Package.appxmanifest.in",
+  "tools/editor/packaging/linux/**",
+  "tools/editor/packaging/macos/AppIcon.icon/Assets/icon_backdrop.svg",
+  "tools/editor/packaging/macos/AppIcon.icon/icon.json",
+  "tools/editor/packaging/macos/dmg-background.svg",
+  "tools/editor/packaging/tile-backdrop-wide.svg",
+  "tools/editor/packaging/tile-backdrop.svg",
+  "tools/editor/packaging/windows-appx/Package.appxmanifest.in",
   "internal/editor-preview/ui/assets/slinterface-*.svg",
   "internal/editor-preview/ui/visual-editor/example/image.png",
   "internal/editor-preview/ui/visual-editor/example/title-frame.png",
