+++ internal/editor-preview/Cargo.toml
@@ -91,9 +91,11 @@ i-slint-backend-testing = { workspace = true, optional = true }
 spin_on = { workspace = true, optional = true }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+directories = "6.0"
 tokio = { version = "1.50.0", features = ["rt"] }
 mdns-sd = { version = "0.20.0", optional = true }
 rfd = { version = "0.17.2", optional = true }
+tempfile = "3"
 
 # Used by the visual editor for the macOS unified title bar.
 [target.'cfg(target_os = "macos")'.dependencies]
