+++ internal/editor-preview/settings_store.rs
