+++ internal/editor-preview/ui/visual-editor/AGENTS.md
@@ -12,7 +12,7 @@ Run the app with Skia against the simple local fixture:
 ```sh
 SLINT_ENABLE_EXPERIMENTAL_FEATURES=1 \
 SLINT_BACKEND=winit-skia \
-cargo run -p slint-lsp --example slint-editor \
+cargo run -p slint-editor \
   --no-default-features \
   --features backend-winit,renderer-skia \
   -- internal/editor-preview/ui/visual-editor/example/simple_preview.slint
@@ -25,9 +25,9 @@ Important:
 - Use Skia. Do not silently switch to the software renderer. If Skia needs a
   first-time `skia-bindings` download, ask for network approval and keep using
   Skia.
-- Use `-p slint-lsp --no-default-features`. The unqualified
-  `cargo run --example slint-editor` route can pull Qt on macOS and fail in
-  sandboxes when `ccache` writes under `~/Library/Caches/ccache`.
+- Use `-p slint-editor --no-default-features`. Dropping
+  `--no-default-features` can pull Qt on macOS and fail in sandboxes when
+  `ccache` writes under `~/Library/Caches/ccache`.
 - `SLINT_ENABLE_EXPERIMENTAL_FEATURES=1` is required because the visual editor
   uses internal/experimental types such as `component-factory`.
 - Use `internal/editor-preview/ui/visual-editor/example/simple_preview.slint` when it exists.
@@ -46,7 +46,7 @@ SLINT_ENABLE_EXPERIMENTAL_FEATURES=1 \
 SLINT_EMIT_DEBUG_INFO=1 \
 SLINT_MCP_PORT=9315 \
 SLINT_BACKEND=winit-skia \
-cargo run -p slint-lsp --example slint-editor \
+cargo run -p slint-editor \
   --no-default-features \
   --features backend-winit,renderer-skia,slint/mcp \
   -- internal/editor-preview/ui/visual-editor/example/simple_preview.slint
@@ -79,7 +79,7 @@ release mode so UI interaction stays responsive:
 SLINT_ENABLE_EXPERIMENTAL_FEATURES=1 \
 SLINT_LIVE_PREVIEW=1 \
 SLINT_BACKEND=winit-skia \
-cargo run --release -p slint-lsp --example slint-editor \
+cargo run --release -p slint-editor \
   --no-default-features \
   --features backend-winit,renderer-skia,slint/live-preview \
   -- internal/editor-preview/ui/visual-editor/example/Main.slint
@@ -94,8 +94,8 @@ switch renderer or app entry point.
 
 ## Architecture
 
-- Entry point: `tools/lsp/editor_main.rs` starts embedded LSP state, then calls
-  `preview::run(..., PreviewUiKind::Editor)`.
+- Entry point: `tools/editor/main.rs` starts embedded LSP state, then creates
+  the editor window with `preview::ui::create_ui(..., PreviewUiKind::Editor)`.
 - UI creation: `internal/editor-preview/preview/ui.rs` creates `EditorUi`, wires the shared
   `Api` global, and registers callbacks.
 - Live preview surface: `EditorCanvas` embeds the compiled target component via
