+++ tools/editor/dev.slint.VisualEditor.yml
@@ -73,17 +73,17 @@ modules:
         SKIA_GN_COMMAND: /app/bin/gn
     build-commands:
       - cargo --offline fetch --manifest-path Cargo.toml
-      - cargo build --manifest-path tools/lsp/Cargo.toml --example slint-editor
+      - cargo build --manifest-path tools/editor/Cargo.toml
         --profile package-release --offline --no-default-features
         --features backend-winit,renderer-skia
-      - install -Dm0755 target/package-release/examples/slint-editor ${FLATPAK_DEST}/bin/slint-visual-editor
-      - install -Dm0644 tools/lsp/packaging/linux/${FLATPAK_ID}.desktop ${FLATPAK_DEST}/share/applications/${FLATPAK_ID}.desktop
+      - install -Dm0755 target/package-release/slint-editor ${FLATPAK_DEST}/bin/slint-visual-editor
+      - install -Dm0644 tools/editor/packaging/linux/${FLATPAK_ID}.desktop ${FLATPAK_DEST}/share/applications/${FLATPAK_ID}.desktop
       # The icon references the shared logo and tile backdrop, and librsvg only
       # resolves references within the document's own directory
-      - cp logo/slint-logo-small-dark.svg tools/lsp/packaging/tile-backdrop.svg tools/lsp/packaging/linux/
+      - cp logo/slint-logo-small-dark.svg tools/editor/packaging/tile-backdrop.svg tools/editor/packaging/linux/
       - >
         for size in 128 256 512; do
-        rsvg-convert -w $size -h $size tools/lsp/packaging/linux/${FLATPAK_ID}.svg -o icon-$size.png || exit 1;
+        rsvg-convert -w $size -h $size tools/editor/packaging/linux/${FLATPAK_ID}.svg -o icon-$size.png || exit 1;
         install -Dm0644 icon-$size.png
         ${FLATPAK_DEST}/share/icons/hicolor/${size}x${size}/apps/${FLATPAK_ID}.png || exit 1;
         done
@@ -91,10 +91,10 @@ modules:
       # version and the commit date of the checkout being built
       - >
         sed -e "s/@VERSION@/$(cargo metadata --no-deps --offline --format-version 1
-        --manifest-path tools/lsp/Cargo.toml |
-        jq -r '.packages[] | select(.name == "slint-lsp") | .version')/"
+        --manifest-path tools/editor/Cargo.toml |
+        jq -r '.packages[] | select(.name == "slint-editor") | .version')/"
         -e "s/@DATE@/$(git log -1 --format=%cs)/"
-        tools/lsp/packaging/linux/${FLATPAK_ID}.metainfo.xml.in > ${FLATPAK_ID}.metainfo.xml
+        tools/editor/packaging/linux/${FLATPAK_ID}.metainfo.xml.in > ${FLATPAK_ID}.metainfo.xml
       - appstreamcli validate --no-net ${FLATPAK_ID}.metainfo.xml
       - install -Dm0644 ${FLATPAK_ID}.metainfo.xml ${FLATPAK_DEST}/share/metainfo/${FLATPAK_ID}.metainfo.xml
     sources:
