+++ tools/editor/packaging/linux/dev.slint.VisualEditor.desktop
