+++ tools/editor/packaging/linux/dev.slint.VisualEditor.metainfo.xml.in
