+++ tools/editor/packaging/macos/AppIcon.icon/icon.json
