+++ tools/editor/packaging/windows-appx/Package.appxmanifest.in
